cdk/requirements.txt
aws-cdk-lib>=2.223.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-bedrock-agentcore-alpha>=2.223.0a0
